Re-arranged code as per review comments
