+++ src/collector/__init__.py
@@ -1,0 +1,1 @@
+"""Google Bigquery data collectors implementation."""
