+++ src/datastore/__init__.py
@@ -1,0 +1,1 @@
+"""Implementation to update data in persistence store like s3 bucket."""
