+++ src/job/__init__.py
@@ -1,0 +1,1 @@
+"""Process pipeline for Google Bigquery data collection and updation."""
